fix: se cambia extension de doubles a float

+++ src/com/metodosnumericos/Calculos.java
@@ -4,25 +4,25 @@ import java.util.Scanner;
 
 public class Calculos {
 
-    Double limiteInferior;
-    Double limiteSuperior;
+    float limiteInferior;
+    float limiteSuperior;
     Integer iteraciones;
-    Double xAnterior = Double.parseDouble("1000");
-    Double xActual;
-    Double error;
+    float xAnterior = Float.parseFloat("1000");
+    float xActual;
+    float error;
 
     public void calcularDatos() {
         System.out.println("Por favor introduzca el valor del limite inferior: ");
         Scanner entradaEscaner = new Scanner(System.in);
-        limiteInferior = Double.parseDouble(entradaEscaner.next());
+        limiteInferior = Float.parseFloat(entradaEscaner.next());
 
         System.out.println("Por favor introduzca el valor del limite superior: ");
         entradaEscaner = new Scanner(System.in);
-        limiteSuperior = Double.parseDouble(entradaEscaner.next());
+        limiteSuperior = Float.parseFloat(entradaEscaner.next());
 
         System.out.println("Por favor introduzca el valor del error: ");
         entradaEscaner = new Scanner(System.in);
-        error = Double.parseDouble(entradaEscaner.next());
+        error = Float.parseFloat(entradaEscaner.next());
 
         xActual = (limiteInferior + limiteSuperior) / 2;
         iteraciones = 0;
@@ -41,7 +41,7 @@ public class Calculos {
         System.out.println("La raiz es: " + xActual + " y el numero de iteraciones fueron: " + iteraciones);
     }
 
-    private Double ejecutarFormula(Double valor) {
-        return Math.pow(valor, 2) - 3 * valor - 4;
+    private float ejecutarFormula(float valor) {
+        return (float) (Math.pow(valor, 2) - (3 * valor) - 4);
     }
 }
